\docType{package}
\name{OLRefclassPkg-package}
\alias{OLRefclassPkg}
\alias{OLRefclassPkg-package}
\title{OLRefclassPkg (title)}
\description{
  A online learning package for reference class documents
  using 'roxygen2' (description)
}
\details{
  This package is implement a online learning algorithm
  (PA, AROW) by reference class methods. (details)
}
\author{
  Author: Yoshiaki AMANO. \cr Maintainer: Yoshiaki AMANO.
  \cr E-Mail: \email{AmanoYoshiaki@next-group.jp}
}
\seealso{
  \code{\link[kernlab:onlearn]{onlearn}}
}
\keyword{documentation}

