% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_OGD}
\alias{run_OGD}
\title{Run online gradient descent}
\usage{
run_OGD(reg, y, X, b_true, learn_rate, PR_ave = TRUE)
}
\arguments{
\item{reg}{run linear \code{reg="linear"} or logistic \code{reg = "logistic"} regression}

\item{y}{response vector}

\item{X}{signal data}

\item{b_true}{true coefficients}

\item{learn_rate}{learning rate}

\item{PR_ave}{logical; for Polyak-Ruppert averaging}
}
\value{
\itemize{
\item \code{b_est} estimated coefficients
\item \code{b_true} true coefficients
\item \code{est_error} estimation error
\item \code{pred_error} prediction error
}
}
\description{
Run online gradient descent
}
\examples{
set.seed(123)
data <- sample_iid_data(reg="linear", n=1000, b_len=10, bn0_len=8)
run_OGD(y=data$y, X=data$X, b_true=data$b, reg="linear", learn_rate=0.01)
}
