% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_RLS}
\alias{run_RLS}
\title{Run recursive least squares}
\usage{
run_RLS(
  Y,
  X,
  initial = 0,
  invxtx = 10^6 * diag(nrow = dim(X)[2], ncol = dim(X)[2]),
  xty = matrix(0, nrow = dim(X)[2], ncol = 1),
  btrue = rep(0, dim(X)[2])
)
}
\arguments{
\item{Y}{response vector}

\item{X}{signal data}

\item{initial}{test}

\item{invxtx}{test}

\item{xty}{test}

\item{btrue}{test}
}
\value{
* `y` response vector
* `X` signal data
* `b` coef vector
}
\description{
Run recursive least squares
}
