% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_iid_data}
\alias{sample_iid_data}
\title{Sample i.i.d data}
\usage{
sample_iid_data(
  reg,
  n = 1000,
  b_len = 500,
  bn0_len = 100,
  signal = 5,
  noise = 1
)
}
\arguments{
\item{reg}{sample data from either [reg="linear"] or [reg = "logistic"]}

\item{n}{sample size}

\item{b_len}{length of coefficient beta vector}

\item{bn0_len}{length of non-zero betas}

\item{signal}{standard deviation of an independent Gaussian distribution with mean 0
from which the non-zero betas are drawn from}

\item{noise}{standard deviation of a normal distribution with mean 0
from which the error vector is drawn from}
}
\value{
test return
}
\description{
Sample i.i.d data
}
