% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_AdaGrad}
\alias{run_AdaGrad}
\title{Run adaptive gradient descent}
\usage{
run_AdaGrad(reg, y, X, b_true, learn_rate, PR_ave = TRUE)
}
\arguments{
\item{reg}{sample data from either `reg="linear"` or `reg = "logistic"`}

\item{y}{response vector}

\item{X}{signal data}

\item{b_true}{true coefficients}

\item{learn_rate}{learning rate}

\item{PR_ave}{logical; for Polyak-Ruppert averaging}
}
\value{
* `y` response vector
* `X` signal data
* `b` coef vector
}
\description{
Run adaptive gradient descent
}
