% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_OGD}
\alias{run_OGD}
\title{Title 1}
\usage{
run_OGD(reg, y, X, b_true, learn_rate, PR_ave = TRUE)
}
\arguments{
\item{reg}{test}

\item{y}{test}

\item{X}{test}

\item{b_true}{test}

\item{learn_rate}{test}

\item{PR_ave}{test}
}
\value{
test
}
\description{
Title 1
}
