% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_RLS}
\alias{run_RLS}
\title{Run recursive least squares}
\usage{
run_RLS(
  Y,
  X,
  initial = 0,
  invxtx = 10^6 * diag(nrow = dim(X)[2], ncol = dim(X)[2]),
  xty = matrix(0, nrow = dim(X)[2], ncol = 1),
  btrue = rep(0, dim(X)[2])
)
}
\arguments{
\item{Y}{response vector}

\item{X}{design matrix}

\item{initial}{how many data points you have used in your previous model}

\item{invxtx}{\eqn{(X^{T}X)^{-1}} from previous model. If there is no previous OLS model
\eqn{10^{6}I_{p}} is used where p is the number of regression parameters}

\item{xty}{matrix \eqn{X^{T}y} from previous OLS model. If there is no previous model
\eqn{0_{p}} is used.}

\item{btrue}{true coefficients}
}
\value{
\itemize{
\item \code{beta} estimated coefficients
\item \code{eeps} estimation error
\item \code{peps} prediction error
}
}
\description{
Run recursive least squares
}
\examples{
testdata = c(1,0,0,1,1,1,1,2,3,1,3,5,1,4,4,1,5,4)
testdata = matrix(testdata,nrow = 6,ncol=3,byrow = TRUE)
response = c(3,5,6,8,10,11)
(beta = run_RLS(response,testdata)[1])
}
