% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_iid_data}
\alias{sample_iid_data}
\title{Sample i.i.d data}
\usage{
sample_iid_data(
  reg,
  n = 1000,
  b_len = 500,
  bn0_len = 100,
  signal = 5,
  noise = 1
)
}
\arguments{
\item{reg}{sample data from either \code{reg="linear"} or \code{reg = "logistic"}}

\item{n}{sample size}

\item{b_len}{length of coefficient vector}

\item{bn0_len}{length of non-zero coefs}

\item{signal}{standard deviation of an independent Gaussian distribution with mean 0
from which the non-zero coefs are drawn from}

\item{noise}{standard deviation of a normal distribution with mean 0
from which the error vector is drawn from}
}
\value{
\itemize{
\item \code{y} response vector
\item \code{X} signal data
\item \code{b} coefficient vector
\item \code{e} residual (for linear regression only)
\item \code{py} probability vector P(y_t = 1| x_t) (for logistic regression only)
}
}
\description{
Sample i.i.d data
}
\examples{
set.seed(123)
data_lr <- sample_iid_data(reg="linear", n=1000, b_len=10, bn0_len=8)
data_lg <- sample_iid_data(reg="logistic", n=1000, b_len=10, bn0_len=8)
}
